move reusable functionality to util module
