+++ urllib3/hsts.py
@@ -1,32 +1,12 @@
-from datetime import datetime, timedelta
 try:
     from itertools import zip_longest
 except ImportError:
     from itertools import izip_longest as zip_longest
-# FIXME will this break on appengine?
-import socket
 
-from urllib3.packages import six
-
-
-split_header_words = six.moves.http_cookiejar.split_header_words
-
+from .util.hsts import (ExpiringRecord, split_domain, is_ipaddress,
+                        parse_max_age, parse_directives_header)
 
 __all__ = ['HSTSManager', 'HSTSStore', 'MemoryHSTSStore']
-
-
-class ExpiringRecord(object):
-    def __init__(self, max_age, _timestamp):
-        self.timestamp = _timestamp or datetime.now()
-        self.max_age = max_age
-
-    @property
-    def end(self):
-        return self.timestamp + timedelta(seconds=self.max_age)
-
-    def is_expired(self, _now=None):
-        now = _now or datetime.now()
-        return self.end < now
 
 
 class HSTSRecord(ExpiringRecord):
@@ -230,39 +210,6 @@ def translate_port(port):
     return port
 
 
-def split_header_word(header):
-    return split_header_words([header])[0]
-
-
-def parse_max_age(string):
-    if string is None:
-        return None
-
-    try:
-        max_age = int(string)
-    except ValueError:
-        return None
-
-    if max_age < 0:
-        return None
-
-    return max_age
-
-
-def parse_directives_header(header):
-    seen_directives = set()
-
-    for k, v in split_header_word(header):
-        k = k.lower()
-
-        if k in seen_directives:
-            yield None, None
-
-        seen_directives.add(k)
-
-        yield k, v
-
-
 def parse_hsts_header(header, domain):
     max_age = None
     include_subdomains = False
@@ -281,30 +228,3 @@ def parse_hsts_header(header, domain):
         return None
 
     return HSTSRecord(domain, max_age, include_subdomains)
-
-
-# FIXME idna?
-def split_domain(domain):
-    return domain.split('.')
-
-
-# FIXME move somewhere else
-def is_ipaddress(domain):
-    return is_v4address(domain) or is_v6address(domain)
-
-
-def check_inet_pton(family, domain):
-    try:
-        socket.inet_pton(family, domain)
-    except socket.error:
-        return False
-
-    return True
-
-
-def is_v6address(domain):
-    return check_inet_pton(socket.AF_INET6, domain)
-
-
-def is_v4address(domain):
-    return check_inet_pton(socket.AF_INET, domain)
